+++ src/hooks/ready/originFixing.js
@@ -5,21 +5,13 @@ async function createActorHook(actor, options, user) {
   // Can't do this in preCreate because the actor id doesn't exist yet.
   if (options.keepId) return;
   if (game.user?.id !== user) return;
-  if (game.modules.get("dnd5e-scriptlets")?.active && game.settings.get("dnd5e-scriptlets", "UpdateCreatedOrigins")) {
-    await OriginFixer.updateActorEffects(actor, true);
-  } else {
-    await OriginFixer.updateActorEffects(actor);
-  }
+  await OriginFixer.updateActorEffects(actor);
 }
 
 async function createTokenHook(tokenDocument, options, user) {
   if (options.keepId) return;
   if (game.user.id !== user) return;
-  if (game.modules.get("dnd5e-scriptlets")?.active && game.settings.get("dnd5e-scriptlets", "UpdateCreatedOrigins")) {
-    await OriginFixer.updateActorEffects(tokenDocument.actor, true);
-  } else {
-    await OriginFixer.updateActorEffects(tokenDocument.actor);
-  }
+  await OriginFixer.updateActorEffects(tokenDocument.actor);
 }
 
 export function setupUpdateCreatedOrigins() {