+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Next Up
 
+# 6.5.7
+
 - Fix issue with removed dnd5e-scriptlets function.
 - Arcane Archer Arrow corrections (2014).
 - Wild Shape and Circle Forms for 2024 Druid. (Thanks to @tminz)